HW-8/kafka-producer/pyproject.toml
[project]
name = "kafka-producer"
version = "0.1.0"
description = ""
authors = [
    {name = "Mykyta Lozhevych",email = "m.lozhevych@setuniversity.edu.ua"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "kafka-python (>=2.2.15,<3.0.0)",
    "kagglehub (>=0.3.12,<0.4.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)"
]

[tool.poetry]
packages = [{include = "kafka_producer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runproducer = "kafka_producer.producer:main"
